git-svn-id: https://svn.code.sf.net/p/ourorgan/svn/trunk@409 69509848-2a73-49d5-a893-c4d8c934025a

+++ src/grandorgue/GOrgueWindchest.h
@@ -41,7 +41,7 @@ public:
 
 	GOrgueWindchest(GrandOrgueFile* organfile);
 	void Load(IniFileConfig& cfg, wxString group);
-	double GetVolume();
+	float GetVolume();
 	unsigned GetTremulantCount();
 	unsigned GetTremulantId(unsigned index);
